Adiciona mensagem destroy
O sistema apresentava o erro ao deletar categoria quando esta continha materiais.

Solução: Apresentar mensagem ao usuário antes de deletar

Fix #22

+++ app/Http/Controllers/CategoriaController.php
@@ -106,6 +106,10 @@ class CategoriaController extends Controller
     public function destroy(Categoria $categoria)
     {
         $this->authorize('admin');
+
+        if ($categoria->materials->isNotEmpty()){
+            return redirect('/categorias')->with('alert-danger', 'Categoria ainda contém materiais. Por favor delete materiais antes!');
+        }
         $categoria->delete();
         return redirect('/categorias');
     }
